src/app/backend/requirements.txt
Flask==3.0.2
Flask-SocketIO
redis
gevent
gevent-websocket
alembic==1.13.1
aniso8601==9.0.1
applicationinsights==0.11.10
astroid==3.0.3
attrs==23.2.0
azure-common==1.1.28
azure-core==1.30.0
azure-identity==1.15.0
azure-keyvault-secrets==4.7.0
bcrypt==4.1.2
blinker==1.7.0
certifi==2024.2.2
cffi==1.16.0
charset-normalizer==3.3.2
click==8.1.7
coverage==7.4.4
cryptography==42.0.4
dill==0.3.8
Flask-Bcrypt==1.0.1
Flask-Cors==4.0.0
Flask-Injector==0.15.0
Flask-JWT-Extended==4.6.0
flask-marshmallow==1.2.0
Flask-Migrate==4.0.5
flask-restx==1.3.0
Flask-SQLAlchemy==3.1.1
greenlet==3.0.3
idna==3.6
importlib_resources==6.3.2
iniconfig==2.0.0
injector==0.21.0
isodate==0.6.1
isort==5.13.2
itsdangerous==2.1.2
Jinja2==3.1.3
jsonschema==4.21.1
jsonschema-specifications==2023.12.1
Mako==1.3.2
MarkupSafe==2.1.5
marshmallow==3.20.2
mccabe==0.7.0
msal==1.26.0
msal-extensions==1.1.0
msrest==0.7.1
oauthlib==3.2.2
packaging==23.2
platformdirs==4.2.0
pluggy==1.4.0
portalocker==2.8.2
pycparser==2.21
PyJWT==2.8.0
pylint==3.0.3
pyodbc==5.1.0
pytest==8.1.1
pytest-cov==5.0.0
pytz==2024.1
referencing==0.34.0
requests==2.31.0
requests-oauthlib==1.3.1
rpds-py==0.18.0
six==1.16.0
SQLAlchemy==2.0.27
tomlkit==0.12.3
typing_extensions==4.9.0
urllib3==2.2.1
Werkzeug==3.0.1
XlsxWriter==3.2.0
zipp==3.17.0
flask-restx==1.3.0
gunicorn==22.0.0
pytest==8.1.1
pytest-cov==5.0.0
